IBlockchainOperations
