+++ src/neoxp/Node/NodeUtility.cs
@@ -1,116 +1,116 @@
-﻿using Akka.Actor;
-using Microsoft.Extensions.Configuration;
-using Neo;
-using Neo.Ledger;
-using Neo.Network.P2P;
-using Neo.Network.P2P.Payloads;
-using Neo.Persistence;
-using Neo.Plugins;
-using Neo.SmartContract;
-using Neo.SmartContract.Manifest;
-using Neo.SmartContract.Native;
-using Neo.Wallets;
-using NeoExpress.Models;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Net;
-using System.Threading;
-using System.Threading.Tasks;
-using static Neo.Ledger.Blockchain;
+﻿// using Akka.Actor;
+// using Microsoft.Extensions.Configuration;
+// using Neo;
+// using Neo.Ledger;
+// using Neo.Network.P2P;
+// using Neo.Network.P2P.Payloads;
+// using Neo.Persistence;
+// using Neo.Plugins;
+// using Neo.SmartContract;
+// using Neo.SmartContract.Manifest;
+// using Neo.SmartContract.Native;
+// using Neo.Wallets;
+// using NeoExpress.Models;
+// using System;
+// using System.Collections.Generic;
+// using System.IO;
+// using System.Linq;
+// using System.Net;
+// using System.Threading;
+// using System.Threading.Tasks;
+// using static Neo.Ledger.Blockchain;
 
-namespace NeoExpress.Node
-{
-    using StackItem = Neo.VM.Types.StackItem;
+// namespace NeoExpress.Node
+// {
+//     using StackItem = Neo.VM.Types.StackItem;
 
-    internal static class NodeUtility
-    {
-        public static bool InitializeProtocolSettings(ExpressChain chain, uint secondsPerBlock = 0)
-        {
-            secondsPerBlock = secondsPerBlock == 0 ? 15 : secondsPerBlock;
+//     internal static class NodeUtility
+//     {
+//         public static bool InitializeProtocolSettings(ExpressChain chain, uint secondsPerBlock = 0)
+//         {
+//             secondsPerBlock = secondsPerBlock == 0 ? 15 : secondsPerBlock;
 
-            IEnumerable<KeyValuePair<string, string>> settings()
-            {
-                yield return new KeyValuePair<string, string>(
-                    "ProtocolConfiguration:Magic", $"{chain.Magic}");
-                yield return new KeyValuePair<string, string>(
-                    "ProtocolConfiguration:MillisecondsPerBlock", $"{secondsPerBlock * 1000}");
-                yield return new KeyValuePair<string, string>(
-                    "ProtocolConfiguration:ValidatorsCount", $"{chain.ConsensusNodes.Count}");
+//             IEnumerable<KeyValuePair<string, string>> settings()
+//             {
+//                 yield return new KeyValuePair<string, string>(
+//                     "ProtocolConfiguration:Magic", $"{chain.Magic}");
+//                 yield return new KeyValuePair<string, string>(
+//                     "ProtocolConfiguration:MillisecondsPerBlock", $"{secondsPerBlock * 1000}");
+//                 yield return new KeyValuePair<string, string>(
+//                     "ProtocolConfiguration:ValidatorsCount", $"{chain.ConsensusNodes.Count}");
 
-                foreach (var (node, index) in chain.ConsensusNodes.Select((n, i) => (n, i)))
-                {
-                    var privateKey = node.Wallet.Accounts
-                        .Select(a => a.PrivateKey)
-                        .Distinct().Single().HexToBytes();
-                    var encodedPublicKey = new KeyPair(privateKey).PublicKey
-                        .EncodePoint(true).ToHexString();
-                    yield return new KeyValuePair<string, string>(
-                        $"ProtocolConfiguration:StandbyCommittee:{index}", encodedPublicKey);
-                    yield return new KeyValuePair<string, string>(
-                        $"ProtocolConfiguration:SeedList:{index}", $"{IPAddress.Loopback}:{node.TcpPort}");
-                }
-            }
+//                 foreach (var (node, index) in chain.ConsensusNodes.Select((n, i) => (n, i)))
+//                 {
+//                     var privateKey = node.Wallet.Accounts
+//                         .Select(a => a.PrivateKey)
+//                         .Distinct().Single().HexToBytes();
+//                     var encodedPublicKey = new KeyPair(privateKey).PublicKey
+//                         .EncodePoint(true).ToHexString();
+//                     yield return new KeyValuePair<string, string>(
+//                         $"ProtocolConfiguration:StandbyCommittee:{index}", encodedPublicKey);
+//                     yield return new KeyValuePair<string, string>(
+//                         $"ProtocolConfiguration:SeedList:{index}", $"{IPAddress.Loopback}:{node.TcpPort}");
+//                 }
+//             }
 
-            var config = new ConfigurationBuilder()
-                .AddInMemoryCollection(settings())
-                .Build();
+//             var config = new ConfigurationBuilder()
+//                 .AddInMemoryCollection(settings())
+//                 .Build();
 
-            return ProtocolSettings.Initialize(config);
-        }
+//             return ProtocolSettings.Initialize(config);
+//         }
 
-        public static Task RunAsync(Func<IStore> getStore, ExpressConsensusNode node, bool enableTrace, TextWriter writer, CancellationToken cancellationToken)
-        {
-            throw new NotImplementedException();
-        }
+//         public static Task RunAsync(Func<IStore> getStore, ExpressConsensusNode node, bool enableTrace, TextWriter writer, CancellationToken cancellationToken)
+//         {
+//             throw new NotImplementedException();
+//         }
         
-        public static Task RunAsync(IStore store, ExpressConsensusNode node, bool enableTrace, TextWriter writer, CancellationToken cancellationToken)
-        {
-            // writer.WriteLine(store.GetType().Name);
+//         public static Task RunAsync(IStore store, ExpressConsensusNode node, bool enableTrace, TextWriter writer, CancellationToken cancellationToken)
+//         {
+//             // writer.WriteLine(store.GetType().Name);
 
-            var tcs = new TaskCompletionSource<bool>();
+//             var tcs = new TaskCompletionSource<bool>();
 
-            _ = Task.Run(() =>
-            {
-                try
-                {
-                    var wallet = DevWallet.FromExpressWallet(node.Wallet);
-                    var multiSigAccount = node.Wallet.Accounts.Single(a => a.IsMultiSigContract());
+//             _ = Task.Run(() =>
+//             {
+//                 try
+//                 {
+//                     var wallet = DevWallet.FromExpressWallet(node.Wallet);
+//                     var multiSigAccount = node.Wallet.Accounts.Single(a => a.IsMultiSigContract());
 
-                    var logPlugin = new LogPlugin(writer);
-                    var storageProvider = new ExpressStorageProvider(store);
-                    var appEngineProvider = enableTrace ? new ExpressApplicationEngineProvider() : null;
-                    var appLogsPlugin = new ExpressAppLogsPlugin(store);
+//                     var logPlugin = new LogPlugin(writer);
+//                     var storageProvider = new ExpressStorageProvider(store);
+//                     var appEngineProvider = enableTrace ? new ExpressApplicationEngineProvider() : null;
+//                     var appLogsPlugin = new ExpressAppLogsPlugin(store);
 
-                    using var system = new NeoSystem(storageProvider.Name);
-                    var rpcSettings = new Neo.Plugins.RpcServerSettings(port: node.RpcPort);
-                    var rpcServer = new Neo.Plugins.RpcServer(system, rpcSettings);
-                    var expressRpcServer = new ExpressRpcServer(multiSigAccount);
-                    rpcServer.RegisterMethods(expressRpcServer);
-                    rpcServer.RegisterMethods(appLogsPlugin);
-                    rpcServer.StartRpcServer();
+//                     using var system = new NeoSystem(storageProvider.Name);
+//                     var rpcSettings = new Neo.Plugins.RpcServerSettings(port: node.RpcPort);
+//                     var rpcServer = new Neo.Plugins.RpcServer(system, rpcSettings);
+//                     var expressRpcServer = new ExpressRpcServer(multiSigAccount);
+//                     rpcServer.RegisterMethods(expressRpcServer);
+//                     rpcServer.RegisterMethods(appLogsPlugin);
+//                     rpcServer.StartRpcServer();
 
-                    system.StartNode(new ChannelsConfig
-                    {
-                        Tcp = new IPEndPoint(IPAddress.Loopback, node.TcpPort),
-                        WebSocket = new IPEndPoint(IPAddress.Loopback, node.WebSocketPort),
-                    });
-                    system.StartConsensus(wallet);
+//                     system.StartNode(new ChannelsConfig
+//                     {
+//                         Tcp = new IPEndPoint(IPAddress.Loopback, node.TcpPort),
+//                         WebSocket = new IPEndPoint(IPAddress.Loopback, node.WebSocketPort),
+//                     });
+//                     system.StartConsensus(wallet);
 
-                    cancellationToken.WaitHandle.WaitOne();
-                }
-                catch (Exception ex)
-                {
-                    tcs.SetException(ex);
-                }
-                finally
-                {
-                    tcs.TrySetResult(true);
-                }
-            });
+//                     cancellationToken.WaitHandle.WaitOne();
+//                 }
+//                 catch (Exception ex)
+//                 {
+//                     tcs.SetException(ex);
+//                 }
+//                 finally
+//                 {
+//                     tcs.TrySetResult(true);
+//                 }
+//             });
 
-            return tcs.Task;
-        }
-    }
-}
+//             return tcs.Task;
+//         }
+//     }
+// }
