+++ test/test.neoxp/TestChainManager.cs
@@ -12,9 +12,9 @@ namespace test.neo_express
         public void can_resolve_empty_filename()
         {
             var fileSystem = new MockFileSystem();
-            var expected = Path.Join(fileSystem.Directory.GetCurrentDirectory(), ChainManager.DEFAULT_EXPRESS_FILENAME);
+            var expected = Path.Join(fileSystem.Directory.GetCurrentDirectory(), BlockchainOperations.DEFAULT_EXPRESS_FILENAME);
 
-            var chainManager = new ChainManager(fileSystem);
+            var chainManager = new BlockchainOperations(fileSystem);
             var actual = chainManager.ResolveFileName(string.Empty);
 
             actual.Should().Be(expected);
@@ -24,9 +24,9 @@ namespace test.neo_express
         public void can_resolve_provided_filename_without_directory_or_extension()
         {
             var fileSystem = new MockFileSystem();
-            var expected = Path.Join(fileSystem.Directory.GetCurrentDirectory(), "test" + ChainManager.EXPRESS_EXTENSION);
+            var expected = Path.Join(fileSystem.Directory.GetCurrentDirectory(), "test" + BlockchainOperations.EXPRESS_EXTENSION);
 
-            var chainManager = new ChainManager(fileSystem);
+            var chainManager = new BlockchainOperations(fileSystem);
             var actual = chainManager.ResolveFileName("test");
 
             actual.Should().Be(expected);
@@ -36,10 +36,10 @@ namespace test.neo_express
         public void can_resolve_provided_filename_without_directory()
         {
             var fileSystem = new MockFileSystem();
-            var expected = Path.Join(fileSystem.Directory.GetCurrentDirectory(), "test" + ChainManager.EXPRESS_EXTENSION);
+            var expected = Path.Join(fileSystem.Directory.GetCurrentDirectory(), "test" + BlockchainOperations.EXPRESS_EXTENSION);
 
-            var chainManager = new ChainManager(fileSystem);
-            var actual = chainManager.ResolveFileName("test" + ChainManager.EXPRESS_EXTENSION);
+            var chainManager = new BlockchainOperations(fileSystem);
+            var actual = chainManager.ResolveFileName("test" + BlockchainOperations.EXPRESS_EXTENSION);
 
             actual.Should().Be(expected);
         }
@@ -49,9 +49,9 @@ namespace test.neo_express
         public void can_resolve_provided_filename_without_extension()
         {
             var fileSystem = new MockFileSystem();
-            var expected = "x:\\test" + ChainManager.EXPRESS_EXTENSION;
+            var expected = "x:\\test" + BlockchainOperations.EXPRESS_EXTENSION;
 
-            var chainManager = new ChainManager(fileSystem);
+            var chainManager = new BlockchainOperations(fileSystem);
             var actual = chainManager.ResolveFileName("x:\\test");
 
             actual.Should().Be(expected);
@@ -61,9 +61,9 @@ namespace test.neo_express
         public void can_resolve_provided_filename_wrong_extension()
         {
             var fileSystem = new MockFileSystem();
-            var expected = "x:\\test.foo" + ChainManager.EXPRESS_EXTENSION;
+            var expected = "x:\\test.foo" + BlockchainOperations.EXPRESS_EXTENSION;
 
-            var chainManager = new ChainManager(fileSystem);
+            var chainManager = new BlockchainOperations(fileSystem);
             var actual = chainManager.ResolveFileName("x:\\test.foo");
 
             actual.Should().Be(expected);
